InOutMapper.non_spec_based_create into its own codepath
