+++ python_modules/dagster/dagster/_core/definitions/decorators/assets_definition_factory.py
@@ -6,6 +6,7 @@ from typing import (
     Any,
     Callable,
     Dict,
+    Iterable,
     List,
     Mapping,
     NamedTuple,
@@ -18,9 +19,12 @@ from typing import (
 
 import dagster._check as check
 from dagster._core.decorator_utils import get_function_params, get_valid_name_permutations
+from dagster._core.definitions.asset_dep import AssetDep
 from dagster._core.definitions.asset_in import AssetIn
 from dagster._core.definitions.asset_key import AssetKey
+from dagster._core.definitions.asset_out import AssetOut
 from dagster._core.definitions.asset_spec import AssetSpec
+from dagster._core.definitions.assets import ASSET_SUBSET_INPUT_PREFIX
 from dagster._core.definitions.input import In
 from dagster._core.definitions.output import Out
 from dagster._core.definitions.resource_annotation import (
@@ -123,6 +127,42 @@ def build_asset_ins(
     return ins_by_asset_key
 
 
+def build_asset_outs(asset_outs: Mapping[str, AssetOut]) -> Mapping[AssetKey, Tuple[str, Out]]:
+    """Creates a mapping from AssetKey to (name of output, Out object)."""
+    outs_by_asset_key: Dict[AssetKey, Tuple[str, Out]] = {}
+    for output_name, asset_out in asset_outs.items():
+        out = asset_out.to_out()
+        asset_key = asset_out.key or AssetKey(
+            list(filter(None, [*(asset_out.key_prefix or []), output_name]))
+        )
+
+        outs_by_asset_key[asset_key] = (output_name.replace("-", "_"), out)
+
+    return outs_by_asset_key
+
+
+def build_subsettable_asset_ins(
+    asset_ins: Mapping[AssetKey, Tuple[str, In]],
+    asset_outs: Mapping[AssetKey, Tuple[str, Out]],
+    internal_upstream_deps: Iterable[AbstractSet[AssetKey]],
+) -> Mapping[AssetKey, Tuple[str, In]]:
+    """Creates a mapping from AssetKey to (name of input, In object) for any asset key that is not
+    currently an input, but may become one if this asset is subset.
+
+    For example, if a subsettable multi-asset produces both A and C, where C depends on both A and
+    some other asset B, there are some situations where executing just A and C without B will result
+    in these assets being generated by different steps within the same job. In this case, we need
+    a separate input to represent the fact that C depends on A.
+    """
+    # set of asset keys which are upstream of another asset, and are not currently inputs
+    potential_deps = set().union(*internal_upstream_deps).difference(set(asset_ins.keys()))
+    return {
+        key: (f"{ASSET_SUBSET_INPUT_PREFIX}{name}", In(Nothing))
+        for key, (name, _) in asset_outs.items()
+        if key in potential_deps
+    }
+
+
 class InMapping(NamedTuple):
     input_name: str
     input: In
@@ -133,17 +173,26 @@ class OutMapping(NamedTuple):
     output: Out
 
 
+def make_keys_by_output_name(
+    asset_outs: Mapping[AssetKey, Tuple[str, Out]],
+) -> Mapping[str, AssetKey]:
+    return {output_name: asset_key for asset_key, (output_name, _) in asset_outs.items()}
+
+
 class InOutMapper:
     def __init__(
         self,
-        in_mappings: Mapping[AssetKey, InMapping],
-        out_mappings: Mapping[AssetKey, OutMapping],
+        *,
+        asset_ins: Mapping[AssetKey, Tuple[str, In]],
+        asset_outs: Mapping[AssetKey, Tuple[str, Out]],
         check_specs: Sequence[AssetCheckSpec],
+        internal_deps: Mapping[AssetKey, Set[AssetKey]],
         can_subset: bool,
     ) -> None:
-        self.in_mappings = in_mappings
-        self.out_mappings = out_mappings
+        self._passed_asset_ins = asset_ins
+        self.asset_outs = asset_outs
         self.check_specs = check_specs
+        self.internal_deps = internal_deps
         self.can_subset = can_subset
 
     @staticmethod
@@ -192,30 +241,120 @@ class InOutMapper:
             )
         remaining_upstream_keys = {key for key in upstream_keys if key not in loaded_upstreams}
         asset_ins = build_asset_ins(fn, explicit_ins, deps=remaining_upstream_keys)
-        return InOutMapper.from_asset_ins_and_asset_outs(
-            asset_ins, output_tuples_by_asset_key, check_specs, can_subset
+
+        internal_deps = {
+            spec.key: {dep.asset_key for dep in spec.deps}
+            for spec in specs
+            if spec.deps is not None
+        }
+
+        return InOutMapper(
+            asset_ins=asset_ins,
+            asset_outs=output_tuples_by_asset_key,
+            check_specs=check_specs,
+            internal_deps=internal_deps,
+            can_subset=can_subset,
         )
 
     @staticmethod
-    def from_asset_ins_and_asset_outs(
-        asset_ins: Mapping[AssetKey, Tuple[str, In]],
-        asset_outs: Mapping[AssetKey, Tuple[str, Out]],
+    def non_spec_based_create(
+        *,
+        asset_out_map: Mapping[str, AssetOut],
+        asset_deps: Mapping[str, Set[AssetKey]],
+        upstream_asset_deps: Iterable[AssetDep],
+        ins: Mapping[str, AssetIn],
+        op_name: str,
+        fn: Callable[..., Any],
         check_specs: Sequence[AssetCheckSpec],
         can_subset: bool,
     ):
-        in_mappings = {
+        asset_ins = build_asset_ins(
+            fn,
+            ins or {},
+            deps=({dep.asset_key for dep in upstream_asset_deps} if upstream_asset_deps else set()),
+        )
+        output_tuples_by_asset_key = build_asset_outs(asset_out_map)
+
+        # validate that the asset_ins are a subset of the upstream asset_deps.
+        upstream_internal_asset_keys = set().union(*asset_deps.values())
+        asset_in_keys = set(asset_ins.keys())
+        if asset_deps and not asset_in_keys.issubset(upstream_internal_asset_keys):
+            invalid_asset_in_keys = asset_in_keys - upstream_internal_asset_keys
+            check.failed(
+                f"Invalid asset dependencies: `{invalid_asset_in_keys}` specified as asset"
+                " inputs, but are not specified in `internal_asset_deps`. Asset inputs must"
+                " be associated with an output produced by the asset."
+            )
+
+        # validate that the asset_deps make sense
+        valid_asset_deps = asset_in_keys | set(output_tuples_by_asset_key.keys())
+        for out_name, asset_keys in asset_deps.items():
+            if asset_out_map and out_name not in asset_out_map:
+                check.failed(
+                    f"Invalid out key '{out_name}' supplied to `internal_asset_deps` argument"
+                    f" for multi-asset {op_name}. Must be one of the outs for this multi-asset"
+                    f" {list(asset_out_map.keys())[:20]}.",
+                )
+            invalid_asset_deps = asset_keys.difference(valid_asset_deps)
+            check.invariant(
+                not invalid_asset_deps,
+                f"Invalid asset dependencies: {invalid_asset_deps} specified in"
+                f" `internal_asset_deps` argument for multi-asset '{op_name}' on key"
+                f" '{out_name}'. Each specified asset key must be associated with an input to"
+                " the asset or produced by this asset. Valid keys:"
+                f" {list(valid_asset_deps)[:20]}",
+            )
+
+        keys_by_output_name = make_keys_by_output_name(output_tuples_by_asset_key)
+        internal_deps = {keys_by_output_name[name]: asset_deps[name] for name in asset_deps}
+
+        return InOutMapper(
+            asset_ins=asset_ins,
+            asset_outs=output_tuples_by_asset_key,
+            check_specs=check_specs or [],
+            internal_deps=internal_deps,
+            can_subset=can_subset,
+        )
+
+    @cached_property
+    def asset_ins(self) -> Mapping[AssetKey, Tuple[str, In]]:
+        if self.can_subset and self.internal_deps:
+            return {
+                **self._passed_asset_ins,
+                **build_subsettable_asset_ins(
+                    self._passed_asset_ins, self.asset_outs, self.internal_deps.values()
+                ),
+            }
+        else:
+            return self._passed_asset_ins
+
+    @cached_property
+    def in_mappings(self) -> Mapping[AssetKey, InMapping]:
+        return {
             asset_key: InMapping(input_name, in_)
-            for asset_key, (input_name, in_) in asset_ins.items()
+            for asset_key, (input_name, in_) in self.asset_ins.items()
         }
-        out_mappings = {
+
+    @cached_property
+    def out_mappings(self) -> Mapping[AssetKey, OutMapping]:
+        return {
             asset_key: OutMapping(output_name, out_)
-            for asset_key, (output_name, out_) in asset_outs.items()
+            for asset_key, (output_name, out_) in self.asset_outs.items()
         }
-        return InOutMapper(in_mappings, out_mappings, check_specs, can_subset)
 
     @cached_property
     def asset_outs_by_output_name(self) -> Mapping[str, Out]:
         return dict(self.out_mappings.values())
+
+    # @cached_property
+    # def asset_ins_by_input_names(self) -> Mapping[str, In]:
+    #     return dict(self.in_mappings.values())
+
+    @cached_property
+    def asset_keys_by_input_name(self) -> Mapping[str, AssetKey]:
+        return {
+            in_mapping.input_name: asset_key for asset_key, in_mapping in self.in_mappings.items()
+        }
 
     @cached_property
     def asset_keys_by_output_name(self) -> Mapping[str, AssetKey]:
